Add files via upload
This piece of code will use libraries like MTCNN and Keras_VGGFACE along with multiple other standard python libraries and build the model for each person whose memories we wanted to preserve. 
Model would be the embedding vectors of each person's face and this model would be serialized along with the "label" value as a pickle file object.